+++ app/assets/stylesheets/application.scss
@@ -615,6 +615,16 @@ main {
 }
 
 
+
+
+.n-circle {
+  position: absolute;
+  padding-left: 1rem;
+  padding-top: 0rem;
+  color: #efa04c;
+}
+
+
 // ボタン関連
 .button {
   outline:none;
