+++ app/controllers/public/comments_controller.rb
@@ -6,6 +6,7 @@ class Public::CommentsController < ApplicationController
     comment = current_user.comments.new(comment_params)
     comment.post_id = @post.id
     comment.save
+    @post.create_notification_comment!(current_user, comment.id) # 通知
   end
 
   def destroy
