+++ app/controllers/public/favorites_controller.rb
@@ -3,7 +3,8 @@ class Public::FavoritesController < ApplicationController
 
   def create
     @post = Post.find(params[:post_id])
-    favorite = current_user.favorites.new(post_id: @post.id)
+    @post.create_notification_favorite!(current_user)
+    favorite = current_user.favorites.new(post_id: @post.id) # 通知
     favorite.save
   end
