+++ app/controllers/public/relationships_controller.rb
@@ -4,6 +4,7 @@ class Public::RelationshipsController < ApplicationController
   def create
     @user = User.find(params[:user_id])
     current_user.follow(@user)
+    @user.create_notification_follow!(current_user) # 通知
   end
 
   def destroy
