+++ app/views/public/notifications/index.html.erb
@@ -1,0 +1,12 @@
+<h1>通知画面一覧</h1>
+
+<!--自分の投稿に対するいいね、コメントは通知に表示しない-->
+<% notifications = @notifications.where.not(visitor_id: current_user.id) %>
+<% if notifications.exists? %>
+  <%= render notifications %>
+  <%= paginate notifications %>
+<% else %>
+  <p>通知はありません</p>
+<% end %>
+
+
